% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{amp_hist}
\alias{amp_hist}
\title{Amplification histogram}
\usage{
amp_hist()
}
\value{
It will write to jpg file a histogram average amplification rates.
}
\description{
\code{amp_hist} creates a histogram to visualise distribution of average
    amplification rates.
}
\details{
When run it uses the "sample_error_results.csv", an output
    from running \code{\link{gen_errors}}, to produce a histogram to aid in
    visualising the distribution of amplification.

    The plot is printed to screen and also saved as a jpg file to the
    `results/` sub-directory. It can be used to determine an amplification
    threshold for further data cleaning.
}
\examples{
\dontrun{
amp_hist()
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
