% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{amp_plots}
\alias{amp_plots}
\title{Amplification Plots}
\usage{
amp_plots()
}
\value{
It will write to jpg file a panel plot of amplification rates.
}
\description{
\code{amp_plots} produces amplification plots from the sample error results.
}
\details{
When run it uses the sample error results, an output from running
    \code{\link{gen_errors}}, to produce a panel plot to aid in visulising
    amplification rate vs allele error and allelic drop out.

    The plots are printed to screen and also saved as a jpg file to the
    `results/` sub-directory. It can be used to determine an amplification
    threshold for further data cleaning.
}
\examples{
\dontrun{
amp_plots()
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
