% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{amp_splots}
\alias{amp_splots}
\title{Amplification plots}
\usage{
amp_splots()
}
\value{
It will write to jpg file a panel plot of amplification rates.
}
\description{
\code{amp_splots} plots sample amplification rate against sample error rates.
}
\details{
When run it uses the sample error results, an output from running
    \code{\link{gen_errors}}, to produce a panel plot to aid in visualising
    amplification rate vs allele error and allelic drop out.

    The plots are printed to screen and also saved as a jpg file to the
    `results/threshold/` sub-directory. It can be used to determine an amplification
    threshold for further data cleaning. Will error with message if no errors
    in raw data.
}
\examples{
\dontrun{
amp_splots()
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://bartesto.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
