% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readiness.R
\name{amp_threshold}
\alias{amp_threshold}
\title{Applies an amplification threshold and produces interim filtered results.}
\usage{
amp_threshold(at)
}
\arguments{
\item{at}{Numeric value of the average amplification rate to filter the
numerical alleles. Values greater than or equal to are retained.}
}
\value{
It will write to `results/` three csv files:
    \itemize{
        \item filtered alleles data with threshold indicated in the name
        \item sample that were filtered at with threshold indicated in the name
        \item proportion of NA's for the loci}

    It will also print to screen the number of samples filtered out at this
    threshold.
}
\description{
\code{amp_threshold} takes a threshold value and applies it to the numerical
    allele data. Also produces a record of which samples were filtered out and
    proportions of NA's per loci.
}
\details{
The predetermined amplification threshold, ascertained through
    visualisation \code{\link{amp_splots}}, is used to filter the numerical
    alleles data set and written to csv. The names of all the samples that are
    rejected at this threshold are also written to csv file.

    Lastly the proportion of NA's for the loci are written to csv. All outputs
    are written to `results/`.
}
\examples{
\dontrun{
amp_threshold(at = 0.8)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://bartesto.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
