% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{calc_dissimilarity}
\alias{calc_dissimilarity}
\title{Calculate dissimilarity matrix}
\usage{
calc_dissimilarity(slist1)
}
\arguments{
\item{slist1}{A list previously compiled by running \code{\link{calc_mismatches}}}
}
\value{
An updated list containing calculated dissimilarity matrix.
}
\description{
\code{calc_dissimilarity} calculates dissimilarity matrix and stores in the
    input list.
}
\details{
When run it calculates the dissimilarity matrix and writes the
    results back into the input list.
}
\examples{
\dontrun{
calc_mismatches(slist = data_list)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
