% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{calc_mismatches}
\alias{calc_mismatches}
\title{Calculate mismatches}
\usage{
calc_mismatches(slist)
}
\arguments{
\item{slist}{A list previously compiled by running \code{\link{ingest}}}
}
\value{
An updated list containing calculated mismatches between samples.
}
\description{
\code{calc_mismatches} calculates mismatches between samples and stores
    results in the input list.
}
\details{
When run it calculates the mismatch between all combinations of
    samples which have been stored in the input list. It then writes the
    results back into the input list.
}
\examples{
\dontrun{
calc_mismatches(slist = data_list)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
