% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readiness.R
\name{data_in}
\alias{data_in}
\title{Ingest of raw data and creation of two tidy data sets.}
\usage{
data_in(filename, replicates = TRUE, suffix = "_dup")
}
\arguments{
\item{filename}{File name of raw data input csv as a character string.}

\item{replicates}{Logical. TRUE if replicates are present in the raw data.}

\item{suffix}{File suffix to denote sample replicates. Default is the character
string"_dup", change if required. Ignore if replicates == TRUE.}
}
\value{
When assigned to an object it will create a list containing two data
    frames.
}
\description{
\code{data_in} reads in raw data SNP genotype results file and creates two
    clean data sets, one with explicit NA's and the other with replacement
    blank cells.
}
\details{
When run, it reads in a raw data csv file from the `source/` sub-directory
    performs some data cleaning and standardisation, and stores two data frames
    in a list. One data frame contains explicit NA's whilst the other has NA's
    coded as blank cells.

    The raw data expected is SNP genotypes generated through the Agena
    Bioscience MassARRAY genotyping system.
}
\examples{
\dontrun{
data_in(filename = "CAGRF20021407_raw.csv", suffix = "_dup")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
