% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Calculate mismatches and dissimilarity matrix}
\usage{
dissimilarity(filtered_alleles, errors)
}
\arguments{
\item{filtered_alleles}{File name of the filtered numerical allele data csv
as a character string.}

\item{errors}{File name of the sample error results data csv as a character
string.}
}
\value{
An updated list containing calculated mismatches between samples and
    calculated dissimilarity matrix. It will also contain distance objects
    that will aid in visualising group memberships.
}
\description{
\code{dissimilarity} calculates mismatches and dissimilarity matrix between
    samples from filtered data.
}
\details{
When run it takes filtered data and calculates mismatches between
    samples and a dissimilarity matrix. Results are stored in a list.
    Comparing samples for the mismatches takes time and progress status is
    printed to screen. Distance objects are also created and can be used to
    create visualisations to aid in determining paramaters for group
    membership.
}
\examples{
\dontrun{
dist <- dissimilarity(filtered_alleles = "numerical_alleles_filtered_at0.8_mt0.2.csv",
    errors = "sample_error_results.csv")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
