% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{elbow_plot}
\alias{elbow_plot}
\title{Elbow plot}
\usage{
elbow_plot(dist, maxh = 10)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}

\item{maxh}{is the maximum "height", or in this case allowable mismatches,
as an integer and impacts the length of the x-axis. Default is 10.}
}
\value{
It will write to jpg file an elbow plot.
}
\description{
\code{elbow_plot} creates an elbow plot.
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}} and generates an elbow plot. An elbow plot can
    aid in visualising different "height" parameters and corresponding number
    of groups created.  Here the "height" parameter equates to number of
    mismatches. The plot is written to the `results/cluster/`
    sub-directory.
}
\examples{
\dontrun{
elbow_plot(dist = dissimilarity_list, maxh = 5)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://bartesto.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
