% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{group_membership}
\alias{group_membership}
\title{Produces a numerical allele data set with groups as defined by selected (h) value.}
\usage{
group_membership(dist, h)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}

\item{h}{Integer. Represents determined mismatch value.}
}
\value{
A data frame object of numerical alleles with group assignation to be
    used in further processing also a csv file ("...withGroups.csv") written
    to the `results/cluster/` sub-directory.
}
\description{
\code{group_membership} creates a data frame of numerical alleles based on a
    provided h value.
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}} and a maximum number of allowable mismatches,
    or h value, to assign group membership to each sample.
}
\examples{
\dontrun{
group_data <- group_membership(dist = dissimilarity_list, h = 5)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://bartesto.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
