% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{heat_plot}
\alias{heat_plot}
\title{Mismatch heatplot}
\usage{
heat_plot(dist)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}
}
\value{
It will save and print to screen a plotly interactive html heat map.
}
\description{
\code{heat_plot} creates a heat plot based on mismatches.
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}} and creates a heat map. The heat map can
    help visualise groupings of samples based on mismatch values. It generates
    a "plotly" visualisation in the viewer when using RStudio whereby values
    can be interrogated by hovering the mouse cursor. The heat map can be saved
    to file either by using the interactive menu in the "plotly" html or by using
    the export button in the RStudio viewer. The html version is saved to
    `results/cluster/`.
}
\examples{
\dontrun{
heat_plot(dist = dissimilarity_list)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://bartesto.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
