% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{ingest}
\alias{ingest}
\title{Ingest of filtered data to list container}
\usage{
ingest(filtered_alleles, errors)
}
\arguments{
\item{filtered_alleles}{File name of the filtered numerical allele data csv
as a character string.}

\item{errors}{File name of the sample error results data csv as a character
string.}
}
\value{
A list holding 4 data sets. The filtered numerical alleles, the sample
    errors, a data frame of sample combinations and a matrix ready for
    dissimilarity calculations.
}
\description{
\code{ingest} reads in the filtered numerical alleles and sample errors,
    stores them in a list and converts the same into other formats for
    further processing.
}
\details{
When run it will read in the filtered numerical alleles from the
    `results/threshold/` directory and sample errors data from the `results/`
    directory. It then stores them in a list that will be accessed by
    functions downstream. It also creates a data frame of all possible
    combinations of samples and a matrix that will be used for calculating
    dissimilarity. These are also stored in the list.
}
\examples{
\dontrun{
data_list <- ingest(filtered_alleles = "numerical_alleles_filtered_at0.8_mt0.2.csv",
   errors = "sample_error_results.csv")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
