% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readiness.R
\name{main_errors}
\alias{main_errors}
\title{Calculates genotyping error rates}
\usage{
main_errors(dl)
}
\arguments{
\item{dl}{data list. This is a list containing two data frames as output from
the \code{\link{data_in}} function.}
}
\value{
It will write to csv file genetic error file csv's for further
    processing and/or evaluation.
}
\description{
\code{main_errors} takes the list generated by \code{\link{data_in}} and
    generates genotyping error outputs which are written to csv file.
}
\details{
When run it calculates genotyping error rates (sample amplification
    rate and, when replicate samples provided, allelic dropout and false allele
    rates) on the complete raw data set which it receives in the form of a
    list from the \code{\link{data_in}} function. Outputs are written to csv
    file to the `results/` sub-directory as:
    \itemize{
        \item numerical_alleles.csv
        \item sample_error_results.csv
        \item summary_error_results.csv}

     These outputs are used in downstream processes and are further refined.
}
\examples{
\dontrun{
main_errors(dl)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
