% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{majorities_html}
\alias{majorities_html}
\title{Creates a formatted html table of the majorities data export.}
\usage{
majorities_html(majorities_csv)
}
\arguments{
\item{majorities_csv}{Character vector. Name of the majorities and ties csv.}
}
\value{
An html table of numerical alleles with group assignation, majority
    vote and indicated ties written to the `results/cluster/` sub-directory.
}
\description{
\code{majorities_html} turns the majorities export csv into a formatted html
    table for easy querying.
}
\details{
Takes the csv output from \code{\link{majorities}} and re-formats it
    to an html table. Any ties can be queried thereby guiding manual edits of
    the final groups data.
}
\examples{
\dontrun{
majorities_html(majorities_csv = "hclust_numerical_mismatch_4_group_majorities_and_ties.csv")}

 @author Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}

}
