% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{misassign}
\alias{misassign}
\title{Misassign probability plots and table}
\usage{
misassign(dist, maxh = 10, lt = 0.005, ut = 0.995)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}

\item{maxh}{Integer. It is the maximum "height", or in this case allowable
mismatches, and impacts the length of the x-axis. Default is 10.}

\item{lt}{Numeric. Lower threshold. Represents the lower tail of the between
individuals distribution. Defaults to 0.005.}

\item{ut}{Numeric. Upper threshold. Represents the upper tail of the within
individuals distribution. Defaults to 0.995.}
}
\value{
A number of histogram plots, governed by maxh parameter, showing
    possible overlap and probability of misassignment are written to jpg file.
    A csv summary of key values is also saved.
}
\description{
\code{misassign} compares different values of h and effect
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}} and a maximum number of allowable mismatches
    to generate a series of plots to visualise the two distributions of inclusion
    status to a group (in or out) based on number of loci at different mismatch
    thresholds.

    Misassignment is defined as occurring when the 99.5% quantile of group
    inclusion (red dash line in plots) exceeds the 0.05% quantile of non-inclusion
}
\examples{
\dontrun{
misassign(dist = dissimilarity_list, maxh = 5)
}

}
\author{
Rujiporn Sun, \email{rujiporn.sun@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
