% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readiness.R
\name{miss_threshold}
\alias{miss_threshold}
\title{Applies quality filtering of loci by applying a maximum amount of missing data
allowed per locus above which loci are discarded}
\usage{
miss_threshold(mt, at)
}
\arguments{
\item{mt}{Numeric value of the "missingness" threshold to further filter the
numerical alleles. Values less than the threshold are retained.}

\item{at}{Numeric value of the average amplification threshold used on the
input data. NOTE used to select a pre-made data.}
}
\value{
Writes a numerical alleles data csv, filtered to the `at` and `mt`
    thresholds, to the `results/` sub-directory. It will also print to screen
    the number of loci that will be filtered out at this "missingness" threshold.
}
\description{
\code{miss_threshold} takes a "missingness" threshold and applies it to a
    filtered numerical allele data set and then writes the results to csv.
}
\details{
The predetermined "missingness" threshold, ascertained through
    visualisation \code{\link{miss_hist}}, is used to further filter the
    numerical alleles data set and written to csv to the `results/`
    sub-directory.

    The amplification threshold parameter is used to select the correct
    data set from the `results/` sub-directory that had been created by the
    \code{\link{amp_threshold}} function.
}
\examples{
\dontrun{
miss_threshold(mt = 0.2, at = 0.8)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
