% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{structure_format}
\alias{structure_format}
\title{Converts majorities output to structure format}
\usage{
structure_format(majorities_csv)
}
\arguments{
\item{majorities_csv}{Character vector. Name of the majorities and ties csv.}
}
\value{
a csv in "structure" format of the majority data.
}
\description{
\code{structure_format} converts majorities output to structure format.
}
\details{
Takes the majorities and ties csv, from running
    \code{\link{majorities}} and converts to a "structure" format which is
    saved as a csv to the `results/finalised/` sub-directory.
}
\examples{
\dontrun{
structure_format(majorities_csv = "hclust_numerical_mismatch_4_group_majorities_and_ties.csv")}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
