% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{summary_tables}
\alias{summary_tables}
\title{Summary tables with additional site and spatial information}
\usage{
summary_tables(
  groups_csv,
  metadata,
  prefix,
  sample,
  site_ID,
  field_date,
  lat,
  long
)
}
\arguments{
\item{groups_csv}{Character vector. Name of the numerical allele groups csv.}

\item{metadata}{Character vector. Name of the lookup csv which should be
located in the `source/` directory.}

\item{prefix}{Character vector. Sample names in csv outputs often have a prefix
such as "ID_" to avoid starting with a numeral and the lookup data often
wont. Provide the prefix in the calculated data so that it can be matched
to the lookup data.}

\item{sample}{Character vector. Name of the column in the lookup data that
contains the sample IDs.}

\item{site_ID}{Character vector. Name of the column in the lookup data that
contains the site/cave IDs.}

\item{field_date}{Character vector. Name of the column in the lookup data that
contains the sampling date.}

\item{lat}{Character vector. Name of the column in the lookup data that
contains the latitude (WGS84) of the sample collection site.}

\item{long}{Character vector. Name of the column in the lookup data that
contains the longitude (WGS84) of the sample collection site.}
}
\value{
Three csv tables are written to the `results/finalised/` sub-directory
    as described in details above.
}
\description{
\code{summary_tables} joins individuals to site location metadata to create
    summary tables.
}
\details{
Takes the groups csv, from running \code{\link{majorities}} and a
    provided lookup table of metadata per sample and provides three summary
    tables:
    \itemize{
      \item capture history
      \item individual by date and site
      \item site stats
    }

    The user is required to map required fields from the lookup table to the
    function parameters. At a minimum the lookup should contain the
    following:
    \itemize{
      \item sample ID
      \item site ID
      \item collection date
      \item latitude (WGS84)
      \item longitude (WGS84)
    }
}
\examples{
\dontrun{
summary_tables(groups_csv = "hclust_numerical_mismatch_h4_withGroups.csv",
metadata = "lookup.csv", prefix = "ID_", sample = "sample", site_ID = "roost_name",
field_date = "collection_date", lat = "dec_lat", long = "dec_long")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
