% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readiness.R
\name{workspace}
\alias{workspace}
\title{Creates a directory structure for inputs and outputs for a `ScatMatch` workflow.}
\usage{
workspace()
}
\value{
Two sub-directories, `results/` and `source/` will be created in the
    USERs `RStudio` project directory.
}
\description{
\code{workspace} creates a directory structure for use with the `ScatMatch`
    functions.
}
\details{
When run within an `RStudio` project folder structure (highly
    recommended), it will create  `results/` and `source/` sub-directories.
    After running, the USER should place a copy of the raw data input csv file to the
    newly created `source/` sub-directory. All outputs from `ScatMatch` functions
    will be written to the `results/` sub-directory.
}
\examples{
\dontrun{
workspace()
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
