% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{freq_hist}
\alias{freq_hist}
\title{Mismatch histogram}
\usage{
freq_hist(dist, bins = 30)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}

\item{bins}{numeric. Number of bins to split the data. Defaults to 30.}
}
\value{
It will write to png file a histogram of mismatch frequency.
}
\description{
\code{freq_hist} creates a histogram of the frequency of genotype mismatches.
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}} and generates a histogram. The histogram can
    help visualise the frequency of mismatches and can aid in selecting a
    suitable value for creating groups. The plot is written to the
    `results/cluster/` sub-directory.
}
\examples{
\dontrun{
freq_hist(dist = dissimilarity_list)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
