% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readiness.R
\name{gen_errors}
\alias{gen_errors}
\title{Main function wrapper to create errors from genotype inputs}
\usage{
gen_errors(filename, suffix = "_dup")
}
\arguments{
\item{filename}{File name of raw data input csv as a character string.}

\item{suffix}{File suffix to denote sample replicates. Default is the character
string"_dup", change if required. Ignore if replicates == TRUE.}
}
\value{
It will write to csv file genetic error file csv's for further
    processing and/or evaluation.
}
\description{
\code{gen_errors} takes a raw data file and generates genetic error outputs
    which are written to csv file.
}
\details{
When run it calculates genetic errors on the complete raw data set
    which it reads in from the `source/` sub-directory. The raw data expected
    is SNP genotypes generated through the Agena Bioscience MassARRAY
    genotyping system.

    Outputs are written to csv file to the `results/` sub-directory as:
    \itemize{
        \item numerical_alleles.csv
        \item sample_error_results.csv
        \item summary_error_results.csv}

     These outputs are used in downstream processes and are further refined.
}
\examples{
\dontrun{
gen_errors(filename = "CAGRF20021407_raw.csv", suffix = "_dup")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
