% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{heat_plot}
\alias{heat_plot}
\title{Mismatch heatplot}
\usage{
heat_plot(dist)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}
}
\value{
It will write to png file a heat map.
}
\description{
\code{heat_plot} creates a heat plot based on mismatches.
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}} and generates a heat map. The heat map can
    help visualise groupings of samples based on mismatch values. The plot is
    written to the `results/cluster/` sub-directory. It also generates a
    "plotly" visualisation in the viewer if using RStudio whereby values can
    be interrogated by hovering the mouse cursor.
}
\examples{
\dontrun{
heat_plot(dist = dissimilarity_list)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
