% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{majorities}
\alias{majorities}
\title{Calculates the group majorities and ties data.}
\usage{
majorities(dist, h, filtered_alleles, errors)
}
\arguments{
\item{dist}{A list object created after running \code{\link{dissimilarity}}.}

\item{h}{Integer. Represents determined mismatch value.}

\item{filtered_alleles}{Character vector. Name of the filtered allele data
set created by running \code{\link{miss_threshold}}.}

\item{errors}{Character vector. The name of the sample errors csv.}
}
\value{
A csv of numerical alleles with group assignation, majority vote and
    indicated ties written to the `results/cluster/` sub-directory.
}
\description{
\code{majorities} creates a csv table of numerical alleles with group
    assignation and majority vote values with indications of any within group
    ties.
}
\details{
When run it uses the list object created by running
    \code{\link{dissimilarity}}, a maximum number of allowable mismatches
    (or h value), the filtered allele data set from running
    \code{\link{miss_threshold}} and the sample errors from running
    \code{\link{gen_errors}} to create a csv. The output contains the numerical
    alleles with group assignation in a wide format which also displays the
    majority vote per group and any ties.
}
\examples{
\dontrun{
majorities(dist = dissimilarity_list, h = 5,
filtered_alleles = "numerical_alleles_filtered_at0.8_mt0.2.csv",
errors = "sample_error_results.csv")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://bartesto.github.io/ScatMatch/index.html}
{the ScatMatch website}
}
