% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{miss_hist}
\alias{miss_hist}
\title{Data "missingness" plot}
\usage{
miss_hist()
}
\value{
It will write to jpg file a histogram of NA proportions for the loci.
}
\description{
\code{miss_hist} creates a histogram to visualise NA proportions for the loci.
}
\details{
When run it uses the "loci_NA.csv", an output
    from running \code{\link{amp_threshold}}, to produce a histogram to aid in
    visualising data "missingness".

    The plot is printed to screen and also saved as a jpg file to the
    `results/` sub-directory. It can be used to determine a "missingness"
    threshold for further data cleaning.
}
\examples{
\dontrun{
miss_hist()
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
