% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{miss_threshold}
\alias{miss_threshold}
\title{Applies a "missingness" threshold to filtered numerical allele data.}
\usage{
miss_threshold(mt, at)
}
\arguments{
\item{mt}{Numeric value of the "missingness" threshold to further filter the
numerical alleles. Values less than the threshold are retained.}

\item{at}{Numeric value of the average amplification threshold used on the
input data. NOTE used to select a pre-made data.}
}
\value{
Writes a numerical alleles data csv, filtered to the `at` and `mt`
    thresholds, to the `results/` sub-directory.
}
\description{
\code{miss_threshold} takes a "missingness" threshold and applies it to a
    filtered numerical allele data set. Writes results to csv.
}
\details{
The predetermined "missingness" threshold, ascertained through
    visualisation \code{\link{miss_plot}}, is used to further filter the
    numerical alleles data set and written to csv to the `results/`
    sub-directory.

    The amplification threshold parameter is used to select the correct
    data set from the `results/` sub-directory that had been created by the
    \code{\link{amp_threshold}} function.
}
\examples{
\dontrun{
miss_threshold(mt = 0.2, at = 0.8)
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see  \url{https://dbca-wa.github.io/DBCAscatR/index.html}
{the DBCAscatR website}
}
